Refactored NotificationsAdapter, moved ViewHolder's to separate classes, optimised code
